Add files via upload
this is the original code for the viewing of current orders...however, the user who is logged in to the system, is the only one who can enter and view the current orders. I modified this code, and the cashier is now able to view the orders entered by the waitstaff.